Split power spectrum and bispectrum into separate files
